+++ ChangeLog.md
@@ -3,6 +3,7 @@
 
 - Obfuscation supports removing all captions
 - Properties dialog for graph properties added
+- "FastRenderingNodeCountLimit" optimization removed, seems no longer be needed
 
 ## 8.2 - 2025-02-26